backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

celery>=5.3
django-celery-beat>=2.8
redis==4.6.0     
kombu>=5.3       
psycopg2-binary>=2.9
python-dotenv>=1.0
django-templated-mail>=1.1
gunicorn>=22.0
google-auth>=2.43
requests>=2.32
Pillow>=12.0
cloudinary>=1.44.1 
django-cloudinary-storage>=0.3.0
django-filter>=25
channels
daphne>=4.2
channels>=4.3
channels-redis>=4.3
razorpay>=2



